project/requirements.txt
fastapi[all]==0.75.1
sqlmodel==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.1.2
python-dotenv==0.20.0
pytest-cov==3.0.0
sqlalchemy==1.4.34
pyhumps==3.7.1
pycln==1.3.3
black==22.3.0
